+++ web/apps/photos/src/components/WatchFolder.tsx
@@ -1,4 +1,5 @@
 import { CenteredFill, SpacedRow } from "@/base/components/containers";
+import { DialogCloseIconButton } from "@/base/components/mui/DialogCloseIconButton";
 import { FocusVisibleButton } from "@/base/components/mui/FocusVisibleButton";
 import {
     OverflowMenu,
@@ -14,7 +15,6 @@ import { ensureElectron } from "@/base/electron";
 import { basename, dirname } from "@/base/file-name";
 import type { CollectionMapping, FolderWatch } from "@/base/types/ipc";
 import { CollectionMappingChoice } from "@/new/photos/components/CollectionMappingChoice";
-import { DialogCloseIconButton } from "@/new/photos/components/mui/Dialog";
 import AddIcon from "@mui/icons-material/Add";
 import CheckIcon from "@mui/icons-material/Check";
 import DoNotDisturbOutlinedIcon from "@mui/icons-material/DoNotDisturbOutlined";
