+++ web/packages/new/photos/components/CollectionMappingChoice.tsx
@@ -1,4 +1,5 @@
 import { SpacedRow } from "@/base/components/containers";
+import { DialogCloseIconButton } from "@/base/components/mui/DialogCloseIconButton";
 import { FocusVisibleButton } from "@/base/components/mui/FocusVisibleButton";
 import type { ModalVisibilityProps } from "@/base/components/utils/modal";
 import type { CollectionMapping } from "@/base/types/ipc";
@@ -13,7 +14,6 @@ import {
 } from "@mui/material";
 import { t } from "i18next";
 import React from "react";
-import { DialogCloseIconButton } from "./mui/Dialog";
 
 type CollectionMappingChoiceProps = ModalVisibilityProps & {
     /**
